requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
pydantic-settings==2.2.1
python-dotenv==1.0.1

SQLAlchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1

passlib==1.7.4
bcrypt==3.2.2

python-jose[cryptography]==3.3.0
